feat(desktop): Claude-desktop-style single-row chrome [REL-19]
The main window was already frameless on Windows/Linux, but chrome was
two stacked rows: a mostly-empty 36px TitleBar strip (window controls
only) above the 44px TopBar. Now one row, like Claude Desktop:

- TitleBar.tsx -> WindowControls.tsx: just the min/max/close cluster,
  rendered inline at the right edge of the TopBar, flush to the corner
- TopBar doubles as the title bar: empty areas drag the window,
  double-click toggles maximize
- AuthGate gains window controls (signed-out previously had none)
- macOS unchanged (native traffic lights)
